requirements.txt
numpy==2.2.4
opencv_python==4.11.0.86
pandas==2.2.3
Pillow==11.1.0
scikit_learn==1.6.1
streamlit==1.40.2
torch==2.2.2
tqdm==4.66.5
ultralytics==8.3.97
